refactor(profile): M1 — split suspension_gate_parts.dart (361→158 lines)
Split 7-widget file (361 lines, violating CLAUDE.md §2.1 200-line limit for
widget files) into 3 focused files:

  suspension_gate_parts.dart    (158 lines) — orchestrator + header + reason card
  suspension_gate_status.dart   (172 lines) — countdown/permanent chips, receipt, upheld body
  suspension_gate_cta.dart       (55 lines) — CTA row

Parts file re-exports the other two so existing callers need no import changes.
New test files: suspension_gate_status_test.dart, suspension_gate_cta_test.dart.

Addresses @mahmutkaya's M1 finding at
lib/features/profile/presentation/widgets/suspension_gate_parts.dart
